refactor: some cleaning
