+++ library/ui/src/commonMain/kotlin/com/mr0xf00/easycrop/ui/Controls.kt
@@ -2,14 +2,32 @@ package com.mr0xf00.easycrop.ui
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.*
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.LocalContentColor
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.material.contentColorFor
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.*
-import androidx.compose.runtime.*
+import androidx.compose.material.icons.filled.Lock
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.setValue
+import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithCache
@@ -94,7 +112,7 @@ fun CropperControls(
 }
 
 @Composable
-private fun ButtonsBar(
+fun ButtonsBar(
     modifier: Modifier = Modifier,
     buttons: @Composable () -> Unit
 ) {
@@ -122,7 +140,7 @@ private fun ButtonsBar(
 
 
 @Composable
-private fun ShapeSelectionMenu(
+fun ShapeSelectionMenu(
     onDismiss: () -> Unit,
     shapes: List<CropShape>,
     selected: CropShape,
@@ -137,7 +155,7 @@ private fun ShapeSelectionMenu(
 
 
 @Composable
-private fun ShapeItem(
+fun ShapeItem(
     shape: CropShape, selected: Boolean, onSelect: () -> Unit,
     modifier: Modifier = Modifier
 ) {
@@ -164,7 +182,7 @@ private fun ShapeItem(
 
 
 @Composable
-private fun AspectSelectionMenu(
+fun AspectSelectionMenu(
     onDismiss: () -> Unit,
     region: Rect,
     onRegion: (Rect) -> Unit,
