+++ sample/composeApp/src/commonMain/kotlin/com/mr0xf00/easycrop/sample/picker/ImagePicker.kt
@@ -1,4 +1,4 @@
-package com.mr0xf00.easycrop.ui
+package com.mr0xf00.easycrop.sample.picker
 
 import androidx.compose.runtime.Composable
 import com.mr0xf00.easycrop.core.images.ImageSrc
