+++ sample/composeApp/src/iosMain/kotlin/com/mr0xf00/easycrop/sample/picker/ImagePicker.ios.kt
@@ -1,4 +1,4 @@
-package com.mr0xf00.easycrop.ui
+package com.mr0xf00.easycrop.sample.picker
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
