+++ sample/composeApp/src/iosMain/kotlin/com/mr0xf00/easycrop/sample/ui/MainViewController.kt
@@ -1,7 +1,6 @@
-package com.mr0xf00.easycrop.ui
+package com.mr0xf00.easycrop.sample.ui
 
 import androidx.compose.ui.window.ComposeUIViewController
 import com.mr0xf00.easycrop.sample.presentation.ImagesViewModel
-import com.mr0xf00.easycrop.sample.ui.App
 
 fun MainViewController() = ComposeUIViewController { App(ImagesViewModel()) }
